Refactor CameraView on AnalyzerScreen to Common Component File
